refactor: team module refactor
